Revert "Support vs code lm api"
